Add files via upload
The script is used to find the centerline of lane markers by thresholding the image using HSV, and actually, only V is used.
I still think there is space for improvement, now the centerline is more like a curve than a straight line, this needs to be fixed. Converting pixels to LLA coordinates is another problem, I guess it results from the wrong latitude and longitude limits. Apart from that, different colors of pavement will cause failure, and I do not have a good idea to solve this.